+++ zerver/tests/test_messages.py
@@ -12,7 +12,6 @@ from analytics.models import RealmCount
 from zerver.decorator import JsonableError
 from zerver.lib.actions import (
     check_message,
-    do_add_alert_words,
     do_change_stream_invite_only,
     do_claim_attachments,
     do_create_user,
@@ -34,21 +33,13 @@ from zerver.lib.message import (
     sew_messages_and_reactions,
     update_first_visible_message_id,
 )
-from zerver.lib.soft_deactivation import (
-    add_missing_messages,
-    do_soft_activate_users,
-    do_soft_deactivate_users,
-    reactivate_user_if_soft_deactivated,
-)
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import (
-    get_subscription,
     get_user_messages,
     make_client,
     message_stream_count,
     most_recent_message,
     most_recent_usermessage,
-    queries_captured,
 )
 from zerver.lib.topic import DB_TOPIC_NAME
 from zerver.lib.upload import create_attachment
@@ -58,10 +49,8 @@ from zerver.models import (
     Message,
     Reaction,
     Realm,
-    RealmAuditLog,
     RealmDomain,
     Recipient,
-    Stream,
     Subscription,
     UserMessage,
     UserPresence,
@@ -1050,393 +1039,6 @@ class CheckMessageTest(ZulipTestCase):
         test_bot.realm.deactivated = True
         send_rate_limited_pm_notification_to_bot_owner(test_bot, good_realm, content)
         self.assertEqual(test_bot.last_reminder, None)
-
-class SoftDeactivationMessageTest(ZulipTestCase):
-
-    def test_reactivate_user_if_soft_deactivated(self) -> None:
-        recipient_list  = [self.example_user("hamlet"), self.example_user("iago")]
-        for user_profile in recipient_list:
-            self.subscribe(user_profile, "Denmark")
-
-        sender = self.example_user('iago')
-        stream_name = 'Denmark'
-        topic_name = 'foo'
-
-        def last_realm_audit_log_entry(event_type: int) -> RealmAuditLog:
-            return RealmAuditLog.objects.filter(
-                event_type=event_type,
-            ).order_by('-event_time')[0]
-
-        long_term_idle_user = self.example_user('hamlet')
-        # We are sending this message to ensure that long_term_idle_user has
-        # at least one UserMessage row.
-        self.send_stream_message(long_term_idle_user, stream_name)
-        do_soft_deactivate_users([long_term_idle_user])
-
-        message = 'Test Message 1'
-        message_id = self.send_stream_message(sender, stream_name,
-                                              message, topic_name)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        self.assertNotEqual(idle_user_msg_list[-1].content, message)
-        with queries_captured() as queries:
-            reactivate_user_if_soft_deactivated(long_term_idle_user)
-        self.assert_length(queries, 8)
-        self.assertFalse(long_term_idle_user.long_term_idle)
-        self.assertEqual(last_realm_audit_log_entry(
-            RealmAuditLog.USER_SOFT_ACTIVATED).modified_user, long_term_idle_user)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count + 1)
-        self.assertEqual(idle_user_msg_list[-1].content, message)
-        long_term_idle_user.refresh_from_db()
-        self.assertEqual(long_term_idle_user.last_active_message_id, message_id)
-
-    def test_add_missing_messages(self) -> None:
-        recipient_list  = [self.example_user("hamlet"), self.example_user("iago")]
-        for user_profile in recipient_list:
-            self.subscribe(user_profile, "Denmark")
-
-        sender = self.example_user('iago')
-        realm = sender.realm
-        sending_client = make_client(name="test suite")
-        stream_name = 'Denmark'
-        stream = get_stream(stream_name, realm)
-        topic_name = 'foo'
-
-        def send_fake_message(message_content: str, stream: Stream) -> Message:
-            recipient = stream.recipient
-            message = Message(sender = sender,
-                              recipient = recipient,
-                              content = message_content,
-                              date_sent = timezone_now(),
-                              sending_client = sending_client)
-            message.set_topic_name(topic_name)
-            message.save()
-            return message
-
-        long_term_idle_user = self.example_user('hamlet')
-        self.send_stream_message(long_term_idle_user, stream_name)
-        do_soft_deactivate_users([long_term_idle_user])
-
-        # Test that add_missing_messages() in simplest case of adding a
-        # message for which UserMessage row doesn't exist for this user.
-        sent_message = send_fake_message('Test Message 1', stream)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        self.assertNotEqual(idle_user_msg_list[-1], sent_message)
-        with queries_captured() as queries:
-            add_missing_messages(long_term_idle_user)
-        self.assert_length(queries, 6)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count + 1)
-        self.assertEqual(idle_user_msg_list[-1], sent_message)
-        long_term_idle_user.refresh_from_db()
-        self.assertEqual(long_term_idle_user.last_active_message_id, sent_message.id)
-
-        # Test that add_missing_messages() only adds messages that aren't
-        # already present in the UserMessage table. This test works on the
-        # fact that previous test just above this added a message but didn't
-        # updated the last_active_message_id field for the user.
-        sent_message = send_fake_message('Test Message 2', stream)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        self.assertNotEqual(idle_user_msg_list[-1], sent_message)
-        with queries_captured() as queries:
-            add_missing_messages(long_term_idle_user)
-        self.assert_length(queries, 7)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count + 1)
-        self.assertEqual(idle_user_msg_list[-1], sent_message)
-        long_term_idle_user.refresh_from_db()
-        self.assertEqual(long_term_idle_user.last_active_message_id, sent_message.id)
-
-        # Test UserMessage rows are created correctly in case of stream
-        # Subscription was altered by admin while user was away.
-
-        # Test for a public stream.
-        sent_message_list = []
-        sent_message_list.append(send_fake_message('Test Message 3', stream))
-        # Alter subscription to stream.
-        self.unsubscribe(long_term_idle_user, stream_name)
-        send_fake_message('Test Message 4', stream)
-        self.subscribe(long_term_idle_user, stream_name)
-        sent_message_list.append(send_fake_message('Test Message 5', stream))
-        sent_message_list.reverse()
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        for sent_message in sent_message_list:
-            self.assertNotEqual(idle_user_msg_list.pop(), sent_message)
-        with queries_captured() as queries:
-            add_missing_messages(long_term_idle_user)
-        self.assert_length(queries, 6)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count + 2)
-        for sent_message in sent_message_list:
-            self.assertEqual(idle_user_msg_list.pop(), sent_message)
-        long_term_idle_user.refresh_from_db()
-        self.assertEqual(long_term_idle_user.last_active_message_id, sent_message_list[0].id)
-
-        # Test consecutive subscribe/unsubscribe in a public stream
-        sent_message_list = []
-
-        sent_message_list.append(send_fake_message('Test Message 6', stream))
-        # Unsubscribe from stream and then immediately subscribe back again.
-        self.unsubscribe(long_term_idle_user, stream_name)
-        self.subscribe(long_term_idle_user, stream_name)
-        sent_message_list.append(send_fake_message('Test Message 7', stream))
-        # Again unsubscribe from stream and send a message.
-        # This will make sure that if initially in a unsubscribed state
-        # a consecutive subscribe/unsubscribe doesn't misbehave.
-        self.unsubscribe(long_term_idle_user, stream_name)
-        send_fake_message('Test Message 8', stream)
-        # Do a subscribe and unsubscribe immediately.
-        self.subscribe(long_term_idle_user, stream_name)
-        self.unsubscribe(long_term_idle_user, stream_name)
-
-        sent_message_list.reverse()
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        for sent_message in sent_message_list:
-            self.assertNotEqual(idle_user_msg_list.pop(), sent_message)
-        with queries_captured() as queries:
-            add_missing_messages(long_term_idle_user)
-        self.assert_length(queries, 6)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count + 2)
-        for sent_message in sent_message_list:
-            self.assertEqual(idle_user_msg_list.pop(), sent_message)
-        long_term_idle_user.refresh_from_db()
-        self.assertEqual(long_term_idle_user.last_active_message_id, sent_message_list[0].id)
-
-        # Test for when user unsubscribes before soft deactivation
-        # (must reactivate them in order to do this).
-
-        do_soft_activate_users([long_term_idle_user])
-        self.subscribe(long_term_idle_user, stream_name)
-        # Send a real message to update last_active_message_id
-        sent_message_id = self.send_stream_message(
-            sender, stream_name, 'Test Message 9')
-        self.unsubscribe(long_term_idle_user, stream_name)
-        # Soft deactivate and send another message to the unsubscribed stream.
-        do_soft_deactivate_users([long_term_idle_user])
-        send_fake_message('Test Message 10', stream)
-
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        self.assertEqual(idle_user_msg_list[-1].id, sent_message_id)
-        with queries_captured() as queries:
-            add_missing_messages(long_term_idle_user)
-        # There are no streams to fetch missing messages from, so
-        # the Message.objects query will be avoided.
-        self.assert_length(queries, 4)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        # No new UserMessage rows should have been created.
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count)
-
-        # Note: At this point in this test we have long_term_idle_user
-        # unsubscribed from the 'Denmark' stream.
-
-        # Test for a Private Stream.
-        stream_name = "Core"
-        private_stream = self.make_stream('Core', invite_only=True)
-        self.subscribe(self.example_user("iago"), stream_name)
-        sent_message_list = []
-        send_fake_message('Test Message 11', private_stream)
-        self.subscribe(self.example_user("hamlet"), stream_name)
-        sent_message_list.append(send_fake_message('Test Message 12', private_stream))
-        self.unsubscribe(long_term_idle_user, stream_name)
-        send_fake_message('Test Message 13', private_stream)
-        self.subscribe(long_term_idle_user, stream_name)
-        sent_message_list.append(send_fake_message('Test Message 14', private_stream))
-        sent_message_list.reverse()
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        for sent_message in sent_message_list:
-            self.assertNotEqual(idle_user_msg_list.pop(), sent_message)
-        with queries_captured() as queries:
-            add_missing_messages(long_term_idle_user)
-        self.assert_length(queries, 6)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count + 2)
-        for sent_message in sent_message_list:
-            self.assertEqual(idle_user_msg_list.pop(), sent_message)
-        long_term_idle_user.refresh_from_db()
-        self.assertEqual(long_term_idle_user.last_active_message_id, sent_message_list[0].id)
-
-    @mock.patch('zerver.lib.soft_deactivation.BULK_CREATE_BATCH_SIZE', 2)
-    def test_add_missing_messages_pagination(self) -> None:
-        recipient_list  = [self.example_user("hamlet"), self.example_user("iago")]
-        stream_name = 'Denmark'
-        for user_profile in recipient_list:
-            self.subscribe(user_profile, stream_name)
-
-        sender = self.example_user('iago')
-        long_term_idle_user = self.example_user('hamlet')
-        self.send_stream_message(long_term_idle_user, stream_name)
-        do_soft_deactivate_users([long_term_idle_user])
-
-        num_new_messages = 5
-        message_ids = []
-        for _ in range(num_new_messages):
-            message_id = self.send_stream_message(sender, stream_name)
-            message_ids.append(message_id)
-
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        idle_user_msg_count = len(idle_user_msg_list)
-        with queries_captured() as queries:
-            add_missing_messages(long_term_idle_user)
-        self.assert_length(queries, 10)
-        idle_user_msg_list = get_user_messages(long_term_idle_user)
-        self.assertEqual(len(idle_user_msg_list), idle_user_msg_count + num_new_messages)
-        long_term_idle_user.refresh_from_db()
-        self.assertEqual(long_term_idle_user.last_active_message_id, message_ids[-1])
-
-    def test_user_message_filter(self) -> None:
-        # In this test we are basically testing out the logic used out in
-        # do_send_messages() in action.py for filtering the messages for which
-        # UserMessage rows should be created for a soft-deactivated user.
-        recipient_list  = [
-            self.example_user("hamlet"),
-            self.example_user("iago"),
-            self.example_user('cordelia'),
-        ]
-        for user_profile in recipient_list:
-            self.subscribe(user_profile, "Denmark")
-
-        cordelia = self.example_user('cordelia')
-        sender = self.example_user('iago')
-        stream_name = 'Denmark'
-        topic_name = 'foo'
-
-        def send_stream_message(content: str) -> None:
-            self.send_stream_message(sender, stream_name,
-                                     content, topic_name)
-
-        def send_personal_message(content: str) -> None:
-            self.send_personal_message(sender, self.example_user("hamlet"), content)
-
-        long_term_idle_user = self.example_user('hamlet')
-        self.send_stream_message(long_term_idle_user, stream_name)
-        do_soft_deactivate_users([long_term_idle_user])
-
-        def assert_um_count(user: UserProfile, count: int) -> None:
-            user_messages = get_user_messages(user)
-            self.assertEqual(len(user_messages), count)
-
-        def assert_last_um_content(user: UserProfile, content: str, negate: bool=False) -> None:
-            user_messages = get_user_messages(user)
-            if negate:
-                self.assertNotEqual(user_messages[-1].content, content)
-            else:
-                self.assertEqual(user_messages[-1].content, content)
-
-        # Test that sending a message to a stream with soft deactivated user
-        # doesn't end up creating UserMessage row for deactivated user.
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test Message 1'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message, negate=True)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
-
-        # Test that sending a message to a stream with soft deactivated user
-        # and push/email notifications on creates a UserMessage row for the
-        # deactivated user.
-        sub = get_subscription(stream_name, long_term_idle_user)
-        sub.push_notifications = True
-        sub.save()
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test private stream message'
-        send_stream_message(message)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count + 1)
-        assert_last_um_content(long_term_idle_user, message)
-        sub.push_notifications = False
-        sub.save()
-
-        # Test sending a private message to soft deactivated user creates
-        # UserMessage row.
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test PM'
-        send_personal_message(message)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count + 1)
-        assert_last_um_content(long_term_idle_user, message)
-
-        # Test UserMessage row is created while user is deactivated if
-        # user itself is mentioned.
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test @**King Hamlet** mention'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count + 1)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
-
-        # Test UserMessage row is not created while user is deactivated if
-        # anyone is mentioned but the user.
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test @**Cordelia Lear**  mention'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message, negate=True)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
-
-        # Test UserMessage row is created while user is deactivated if
-        # there is a wildcard mention such as @all or @everyone
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test @**all** mention'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count + 1)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
-
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test @**everyone** mention'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count + 1)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
-
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Test @**stream** mention'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count + 1)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
-
-        # Test UserMessage row is not created while user is deactivated if there
-        # is a alert word in message.
-        do_add_alert_words(long_term_idle_user, ['test_alert_word'])
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = 'Testing test_alert_word'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count + 1)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
-
-        # Test UserMessage row is created while user is deactivated if
-        # message is a me message.
-        general_user_msg_count = len(get_user_messages(cordelia))
-        soft_deactivated_user_msg_count = len(get_user_messages(long_term_idle_user))
-        message = '/me says test'
-        send_stream_message(message)
-        assert_last_um_content(long_term_idle_user, message, negate=True)
-        assert_um_count(long_term_idle_user, soft_deactivated_user_msg_count)
-        assert_um_count(cordelia, general_user_msg_count + 1)
-        assert_last_um_content(cordelia, message)
 
 class MessageVisibilityTest(ZulipTestCase):
     def test_update_first_visible_message_id(self) -> None:
